==> fund 2004-2017/08-proportion.R <==
#加入csmar的十大重仓股
topten <- read_excel("C://Users//shenfan//Desktop//data//2//前十大重仓股占基金净值.xlsx")
topten <- as.data.table(topten)
setnames(topten, 1:30, c("code", "name", "2004-06-30", "2004-12-31", "2005-06-30", "2005-12-31", "2006-06-30", "2006-12-31", "2007-06-30", "2007-12-31", "2008-06-30", "2008-12-31", "2009-06-30", "2009-12-31", "2010-06-30", "2010-12-31", "2011-06-30", "2011-12-31", "2012-06-30", "2012-12-31", "2013-06-30", "2013-12-31", "2014-06-30", "2014-12-31", "2015-06-30", "2015-12-31", "2016-06-30", "2016-12-31", "2017-06-30", "2017-12-31"))
topten = melt(topten, id.vars = c("code", "name"), measure.vars = c("2004-06-30", "2004-12-31", "2005-06-30", "2005-12-31", "2006-06-30", "2006-12-31", "2007-06-30", "2007-12-31", "2008-06-30", "2008-12-31", "2009-06-30", "2009-12-31", "2010-06-30", "2010-12-31", "2011-06-30", "2011-12-31", "2012-06-30", "2012-12-31", "2013-06-30", "2013-12-31", "2014-06-30", "2014-12-31", "2015-06-30", "2015-12-31", "2016-06-30", "2016-12-31", "2017-06-30", "2017-12-31"))

setnames(topten, c("value", "variable"), c("top.10", "DateQ"))
topten <- topten[, DateQ := as.Date(as.character(DateQ))
	][, name := NULL
	][order(code, DateQ)
	][, year := year(DateQ)
	][, quarter := quarter(DateQ)
	][quarter == 2, sem := 1
	][quarter == 4, sem := 2
	][,c("code","year","sem","top.10")]

# 加入fund5
load("fund5.RData")
fund.6 <- topten[fund.5, on = .(code, year, sem), nomatch = 0]


#加入csmar的股票市值占基金资产净值比
stockpro <- read_excel("C://Users//shenfan//Desktop//data//2//股票市值占基金资产净值比.xlsx")
stockpro <- as.data.table(stockpro)
setnames(stockpro, 1:30, c("code", "name", "2004-06-30", "2004-12-31", "2005-06-30", "2005-12-31", "2006-06-30", "2006-12-31", "2007-06-30", "2007-12-31", "2008-06-30", "2008-12-31", "2009-06-30", "2009-12-31", "2010-06-30", "2010-12-31", "2011-06-30", "2011-12-31", "2012-06-30", "2012-12-31", "2013-06-30", "2013-12-31", "2014-06-30", "2014-12-31", "2015-06-30", "2015-12-31", "2016-06-30", "2016-12-31", "2017-06-30", "2017-12-31"))
stockpro = melt(stockpro, id.vars = c("code", "name"), measure.vars = c("2004-06-30", "2004-12-31", "2005-06-30", "2005-12-31", "2006-06-30", "2006-12-31", "2007-06-30", "2007-12-31", "2008-06-30", "2008-12-31", "2009-06-30", "2009-12-31", "2010-06-30", "2010-12-31", "2011-06-30", "2011-12-31", "2012-06-30", "2012-12-31", "2013-06-30", "2013-12-31", "2014-06-30", "2014-12-31", "2015-06-30", "2015-12-31", "2016-06-30", "2016-12-31", "2017-06-30", "2017-12-31"))

setnames(stockpro, c("value", "variable"), c("stock.proportion", "DateQ"))
stockpro <- stockpro[, DateQ := as.Date(as.character(DateQ))
	][, name := NULL
	][order(code, DateQ)
	][, year := year(DateQ)
	][, quarter := quarter(DateQ)
	][quarter == 2, sem := 1
	][quarter == 4, sem := 2
	][, c("code", "year", "sem", "stock.proportion")]

# 加入fund5
fund.6 <- stockpro[fund.6, on = .(code, year, sem), nomatch = 0]

save(fund.6, file = "fund6.RData")
